Remove StatusChip component
