--- src/components/StatusChip/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./StatusChip";
