--- src/components/StatusChip/types.ts
@@ -1,6 +1,0 @@
-export enum StatusType {
-  INFO = "info",
-  ERROR = "error",
-  WARNING = "warning",
-  SUCCESS = "success"
-}
